Change Renderer and added new member to Displayable
In Displayable, I've added a flag to check if we should render a object.

In Renderer, I've changed it such that it will render displayable
objects in its list. Objects that wants to be display would have to add
themselves to this list. The list is static, so are the functions,
there's a mixed of static and non-static functions and variable, I'm not
sure if we should have it this way. Initial idea was to have one
renderer that does everything, so should we make everything static?